rename(presentation): presentation package 오타 수정
