+++ src/main/kotlin/com/example/moiza/domain/auth/service/GetAccessTokenService.kt
@@ -2,7 +2,7 @@ package com.example.moiza.domain.auth.service
 
 import com.example.moiza.domain.auth.domain.RefreshToken
 import com.example.moiza.domain.auth.domain.repository.RefreshTokenRepository
-import com.example.moiza.domain.auth.persentation.dto.res.AccessTokenResponse
+import com.example.moiza.domain.auth.presentation.dto.res.AccessTokenResponse
 import com.example.moiza.global.security.jwt.JwtTokenProvider
 import com.teaminsert.homepage.global.security.jwt.exception.ExpiredJwtException
 import org.springframework.data.repository.findByIdOrNull
